requirements.txt
Pillow>=10.0.0
beautifulsoup4>=4.12.0
volcengine-python-sdk[ark]>=1.0.0
requests>=2.31.0
opencv-python>=4.8.0
google-generativeai>=0.3.0
numpy>=1.24.0
openai>=1.0.0
playwright>=1.40.0
scikit-learn>=1.3.0
scipy>=1.11.0
pandas>=2.0.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
keras>=2.15.0
tensorflow>=2.15.0
tqdm>=4.65.0
transformers>=4.40.0
torch>=2.1.0
